requirements.txt
apache-beam==2.51.0
cloudevents==1.10.0
functions-framework==3.4.0
google-api-core==2.11.0
google-auth==2.23.4
google-cloud-core==2.3.2
google-cloud-storage==2.12.0
pyOpenSSL==21.0.0
python-dotenv==1.0.0
google-cloud-bigquery==3.13.0
pandas==2.1.3